Emit pairing_failed event when given wrong secret
To enable pairing code based authentication without having to register a new pairing secret for each new user session, we are emitting the pairing failure event when a wrong secret has been provided. This event can be caught by the chatbot which can in turn authorize the access for the respective user session.