Update device.js
Some pairing message parameters are now wrapped inside a pairing request structure that is provided by the `pairing_attempt` event to its subscribers. This works as a potential custom filter for the subscribers to allow modifying some of the parameters before fully handling the pairing message. For example, subscribers may want to modify the device name or pairing secret before responding to the pairing message.